to long

+++ src/main/java/fr/aresrpg/tofumanchou/domain/data/item/Item.java
@@ -27,7 +27,7 @@ public interface Item {
 
 	Effect[] getEffects();
 
-	int getPrice();
+	long getPrice();
 
 	int getSkin();
